requirement.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
billiard==3.3.0.23
celery==3.1.25
Django==3.1.7
djangorestframework==3.12.4
kombu==3.0.37
Pillow==8.1.2
psycopg2==2.8.6
pytz==2021.1
redis==2.10.6
sqlparse==0.4.1
unicode==2.8
